# documentation/07-CONTRACTS/OPENAPI_SPEC_V2.5.2.yaml
openapi: 3.0.0
info:
  title: TikTrack Project Phoenix API (Master Release)
  version: 2.5.2
  description: Unified API Specification (Fortress Protocol) - Updated with Authentication & Identity endpoints
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
paths:
...
  /users/me:
    get:
      summary: "Get current user profile (D25)"
      tags:
        - users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "User profile"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: "Update user profile (D25)"
      tags:
        - users
      security:
        - bearerAuth: []
      description: "Update user profile fields. Phone number changes require verification via /auth/verify-phone."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: "Profile updated successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: "Invalid input"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/me/password:
    put:
      summary: "Change user password"
      tags:
        - users
      security:
        - bearerAuth: []
      description: |
        Change password for authenticated user.
        Requires verification of current password (old_password).
        Rate limited: 5 attempts per 15 minutes per user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChangeRequest'
      responses:
        '200':
          description: "Password changed successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordChangeResponse'
        '401':
          description: "Invalid password"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: "Rate limit exceeded (5 attempts per 15 minutes)"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /user/api-keys:
    get:
      summary: "List user's API keys (D24)"
      tags:
        - api-keys
      security:
        - bearerAuth: []
      description: "List all API keys for the authenticated user. Returns masked keys."
      responses:
        '200':
          description: "List of API keys"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserApiKeyResponse'
        '401':
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: "Create new API key (D24)"
      tags:
        - api-keys
      security:
        - bearerAuth: []
      description: "Create a new API key for a provider. Keys are encrypted before storage."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserApiKeyCreate'
      responses:
        '201':
          description: "API key created successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserApiKeyResponse'
        '400':
          description: "Invalid input or duplicate key"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /user/api-keys/{key_id}:
    put:
      summary: "Update API key (D24)"
      tags:
        - api-keys
      security:
        - bearerAuth: []
      parameters:
        - name: key_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-7][0-9A-HJKMNP-TV-Z]{25}$'
          description: "API key ULID"
      description: "Update API key (label, status, or re-encrypt keys)."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                provider_label:
                  type: string
                  nullable: true
                is_active:
                  type: boolean
                  nullable: true
                api_key:
                  type: string
                  nullable: true
                api_secret:
                  type: string
                  nullable: true
                additional_config:
                  type: object
                  nullable: true
      responses:
        '200':
          description: "API key updated successfully"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserApiKeyResponse'
        '400':
          description: "Invalid input"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "API key not found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: "Delete API key (D24)"
      tags:
        - api-keys
      security:
        - bearerAuth: []
      parameters:
        - name: key_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-7][0-9A-HJKMNP-TV-Z]{25}$'
          description: "API key ULID"
      description: "Delete API key (soft delete - sets deleted_at timestamp)."
      responses:
        '204':
          description: "API key deleted successfully"
        '401':
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "API key not found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /user/api-keys/{key_id}/verify:
    post:
      summary: "Verify API key against provider (D24)"
      tags:
        - api-keys
      security:
        - bearerAuth: []
      parameters:
        - name: key_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-7][0-9A-HJKMNP-TV-Z]{25}$'
          description: "API key ULID"
      description: "Verify API key against provider API and update verification status."
      responses:
        '200':
          description: "Verification result"
          content:
            application/json:
              schema:
                type: object
                properties:
                  key_id:
                    type: string
                    example: "01ARZ3NDEKTSV4RRFFQ69G5FAV"
                  is_verified:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "API key verified successfully"
        '401':
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "API key not found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token in Authorization header"
  schemas:
...
    UserResponse:
      type: object
      required:
        - external_ulids
        - email
        - created_at
      properties:
        external_ulids:
          type: string
          pattern: '^[0-7][0-9A-HJKMNP-TV-Z]{25}$'
          description: "User ULID (external identifier)"
          example: "01ARZ3NDEKTSV4RRFFQ69G5FAV"
        email:
          type: string
          format: email
          description: "Email address"
          example: "user@example.com"
        phone_numbers:
          type: string
          nullable: true
          description: "Phone number"
          example: "+12025551234"
        user_tier_levels:
          type: string
          enum: [Bronze, Silver, Gold, Platinum]
          default: Bronze
          description: "User tier level"
        username:
          type: string
          nullable: true
          description: "Username"
        display_name:
          type: string
          nullable: true
          description: "Display name"
        role:
          type: string
          enum: [USER, ADMIN, SUPERADMIN]
          default: USER
        is_email_verified:
          type: boolean
          default: false
        phone_verified:
          type: boolean
          default: false
        created_at:
          type: string
          format: date-time
    UserUpdate:
      type: object
      description: "User profile update request schema"
      properties:
        first_name:
          type: string
          maxLength: 100
          nullable: true
          description: "First name"
          example: "John"
        last_name:
          type: string
          maxLength: 100
          nullable: true
          description: "Last name"
          example: "Doe"
        display_name:
          type: string
          maxLength: 100
          nullable: true
          description: "Display name"
          example: "Johnny"
        phone_number:
          type: string
          nullable: true
          pattern: '^\+?[1-9]\d{1,14}$'
          description: "Phone number (E.164 format). Changing phone number resets verification status."
          example: "+1234567890"
        timezone:
          type: string
          maxLength: 50
          nullable: true
          description: "Timezone (e.g., 'America/New_York')"
          example: "America/New_York"
        language:
          type: string
          maxLength: 5
          nullable: true
          description: "Language code (e.g., 'en', 'he')"
          example: "en"
    PasswordChangeRequest:
      type: object
      required:
        - old_password
        - new_password
      description: "Password change request schema for authenticated users"
      properties:
        old_password:
          type: string
          format: password
          minLength: 1
          description: "Current password"
          example: "current_password_123"
        new_password:
          type: string
          format: password
          minLength: 8
          description: "New password (min 8 characters). Must be different from current password."
          example: "new_secure_password_456"
    PasswordChangeResponse:
      type: object
      required:
        - message
      description: "Password change response schema"
      properties:
        message:
          type: string
          description: "Success message"
          example: "Password changed successfully"

    # Password Reset Schemas
...
    UserApiKeyCreate:
      type: object
      required:
        - provider
        - api_key
      properties:
        provider:
          type: string
          enum: [IBKR, POLYGON, YAHOO_FINANCE, ALPHA_VANTAGE, FINNHUB, TWELVE_DATA, IEX_CLOUD, CUSTOM]
          description: "API provider"
          example: "POLYGON"
        provider_label:
          type: string
          maxLength: 100
          nullable: true
          description: "Custom label for this key"
          example: "Production Key"
        api_key:
          type: string
          minLength: 1
          description: "API key (will be encrypted)"
          example: "your_api_key_here"
        api_secret:
          type: string
          nullable: true
          description: "API secret (will be encrypted)"
          example: "your_api_secret_here"
        additional_config:
          type: object
          nullable: true
          description: "Additional configuration (JSON)"
          example: {}
    UserApiKeyResponse:
      type: object
      required:
        - external_ulids
        - provider
        - masked_key
        - is_active
        - is_verified
        - created_at
      properties:
        external_ulids:
          type: string
          pattern: '^[0-7][0-9A-HJKMNP-TV-Z]{25}$'
          description: "API key ULID (external identifier)"
          example: "01ARZ3NDEKTSV4RRFFQ69G5FAV"
        provider:
          type: string
          enum: [IBKR, POLYGON, YAHOO_FINANCE, ALPHA_VANTAGE, FINNHUB, TWELVE_DATA, IEX_CLOUD, CUSTOM]
          description: "API provider"
          example: "POLYGON"
        provider_label:
          type: string
          nullable: true
          description: "Custom label"
          example: "Production Key"
        masked_key:
          type: string
          description: "Masked API key (always returns ********************)"
          example: "********************"
        is_active:
          type: boolean
          description: "Active status"
          example: true
        is_verified:
          type: boolean
          description: "Verification status"
          example: true
        last_verified_at:
          type: string
          format: date-time
          nullable: true
          description: "Last verification timestamp"
          example: "2026-01-31T12:00:00Z"
        created_at:
          type: string
          format: date-time
          description: "Creation timestamp"
          example: "2026-01-31T10:00:00Z"

    # Error Schemas
    ErrorResponse:
      type: object
      required:
        - detail
        - error_code
      properties:
        detail:
          type: string
          description: "Human-readable error message"
          example: "Invalid credentials"
        error_code:
          type: string
          description: "Machine-readable error code for programmatic handling"
          example: "AUTH_INVALID_CREDENTIALS"

    # Trading Schemas (existing)
...